# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Sentinel HTTP API (Express) — health, waitlist, reports, my-reports, pairing, devices, version, and downloads.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pair/session:
    post:
      operationId: createPairSession
      tags: [pair]
      summary: Create a short-lived pair code for the browser
      responses:
        "201":
          description: Pair session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PairSessionCreated"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /pair/session/{code}:
    get:
      operationId: getPairSessionStatus
      tags: [pair]
      summary: Poll pair session until report is ready
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PairSessionPollResponse"
  /pair/push:
    post:
      operationId: pushPairedReport
      tags: [pair]
      summary: Push collector JSON for a pair code (diagnostic script)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PairPushRequest"
      responses:
        "200":
          description: Idempotent — report already linked to this code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PairPushOk"
        "201":
          description: Report created and linked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PairPushCreated"
        "404":
          description: Pair code not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "410":
          description: Pair code expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Payload too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many push attempts for this code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /downloads/latest:
    get:
      operationId: getLatestReleaseJson
      tags: [downloads]
      summary: Latest GitHub release metadata (JSON)
      responses:
        "200":
          description: Release metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LatestReleaseMetadata"
        "502":
          description: Upstream GitHub error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /downloads/latest/{binary}:
    get:
      operationId: redirectLatestBinaryDownload
      tags: [downloads]
      summary: Redirect to the latest release asset for a binary slug
      description: |
        Valid `binary` values — `oneshot`, `setup`, `agent`. Responds with HTTP 302 to the
        GitHub `browser_download_url` when found; otherwise redirects to the releases page.
      parameters:
        - name: binary
          in: path
          required: true
          schema:
            type: string
            enum: [oneshot, setup, agent]
      responses:
        "302":
          description: Redirect to download URL or releases page
          headers:
            Location:
              schema:
                type: string
        "404":
          description: Unknown binary slug
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
      required:
        - error
...
    JsonObject:
      type: object
      additionalProperties: true
...
    PairSessionCreated:
      type: object
      properties:
        code:
          type: string
        expiresAt:
          type: string
          format: date-time
      required:
        - code
        - expiresAt
    PairSessionPollResponse:
      type: object
      properties:
        status:
          type: string
          enum: [pending, ready, expired]
        reportId:
          type: string
        claimToken:
          type: string
      required:
        - status
    PairPushRequest:
      type: object
      properties:
        pairCode:
          type: string
          minLength: 4
          maxLength: 12
        rawJson:
          $ref: "#/components/schemas/JsonObject"
      required:
        - pairCode
        - rawJson
    PairPushOk:
      type: object
      properties:
        ok:
          type: boolean
        reportId:
          type: string
      required:
        - ok
        - reportId
    PairPushCreated:
      type: object
      properties:
        ok:
          type: boolean
        reportId:
          type: string
        claimToken:
          type: string
      required:
        - ok
        - reportId
        - claimToken
...
    LatestReleaseMetadata:
      type: object
      properties:
        version:
          type: string
        name:
          type: string
        publishedAt:
          type: string
          format: date-time
        releaseUrl:
          type: string
          format: uri
        assets:
          type: array
          items:
            $ref: "#/components/schemas/ReleaseAsset"
      required:
        - version
        - name
        - publishedAt
        - releaseUrl
        - assets
    ReleaseAsset:
      type: object
      properties:
        name:
          type: string
        size:
          type: integer
        downloadUrl:
          type: string
          format: uri
        downloads:
          type: integer
      required:
        - name
        - size
        - downloadUrl
        - downloads